requirements.txt
pandas
pandas_ta
matplotlib
MetaTrader5
openai
python-dotenv
